Refactored adjusted timestamps logic
