+++ backend/utils/advanceQueue.js
@@ -9,7 +9,10 @@ async function advanceQueue(locationData) {
         return;
     }
 
-    // // If there are still empty courts, advance the queue further
+    // Keep track of current unique adjustment and the current time
+    currAdjustment = 0;
+    const now = new Date();
+
     queueAdvanced = false;
     for (let i = 0; i < activeFirebaseUIDs.length; i++) {
         
@@ -20,6 +23,11 @@ async function advanceQueue(locationData) {
 
         // If current court is empty, move first player in the current queue to active
         if (activeFirebaseUIDs[i].startsWith('Empty')) {
+            // Create adjusted timestamp for the new player
+            const adjustedTime = new Date(now);
+            adjustedTime.setSeconds(now.getSeconds() + currAdjustment);
+            currAdjustment -= 10;
+            
             // Get UID and nickname of first player in the current queue
             const firstQueuePlayerUID = queueFirebaseUIDs.shift();
             const firstQueuePlayerNickname = queueNicknames.shift();
@@ -27,7 +35,7 @@ async function advanceQueue(locationData) {
             // Update active data to reflect the new player
             activeFirebaseUIDs[i] = firstQueuePlayerUID; 
             activeNicknames[i] = firstQueuePlayerNickname;
-            activeStartTimes[i] = new Date();
+            activeStartTimes[i] = adjustedTime;
             activeWaitingPlayers[i] = false;  
 
             queueAdvanced = true;
